SSAMem: Reorder SSA and SingleBlock opts
Now SingleBlock precedes SSA. This makes sense as singleblock WAW can get
rid of some stores that get in the way of SSA.

Also had to fix SSA and make sure that stores dominate loads. It is
possible to get undef loads which precedes a single store. This causes
uses of ids before their defs.